A new, working version of demog.py replaces demographics.py
